Move all task logic to view, and add a view details button.
